+++ server/C/cem/sedtrans.c
@@ -6,21 +6,21 @@
 #include "BeachGrid.h"
 
 
-double GetTransportVolumePotential(double alpha, double wave_height, double timestep_length);
+float GetTransportVolumePotential(float alpha, float wave_height, float timestep_length);
 void OopsImEmpty(struct BeachGrid* grid, struct BeachNode* node);
 void OopsImFull(struct BeachGrid* grid, struct BeachNode* node);
-double GetDepthOfClosure(struct BeachNode* node, int ref_pos, double shelf_depth_at_ref_pos, double shelf_slope, double shoreface_slope, double shore_angle, double min_shelf_depth_at_closure, int cell_size);
-double ModTowardZero(double a, double m);
-double RoundRadians(double angle, double round_to, double bias);
-double GetDir(double shore_angle);
-struct BeachNode* GetNodeInDir(struct BeachGrid* grid, struct BeachNode* node, double dir);
+float GetDepthOfClosure(struct BeachNode* node, int ref_pos, float shelf_depth_at_ref_pos, float shelf_slope, float shoreface_slope, float shore_angle, float min_shelf_depth_at_closure, int cell_size);
+float ModTowardZero(float a, float m);
+float RoundRadians(float angle, float round_to, float bias);
+float GetDir(float shore_angle);
+struct BeachNode* GetNodeInDir(struct BeachGrid* grid, struct BeachNode* node, float dir);
 
 
-void WaveTransformation(struct BeachGrid* grid, double wave_angle, double wave_period, double wave_height, double timestep_length)
+void WaveTransformation(struct BeachGrid* grid, float wave_angle, float wave_period, float wave_height, float timestep_length)
 {
-	double start_depth = 3 * wave_height;       // (meters) depth to begin refraction calculations
-	double refract_step = 0.2;                  // (meters) step size to iterate through depth
-	double k_break = 0.5;                       // coefficient such that waves break at Hs > k_break*depth
+	float start_depth = 3 * wave_height;       // (meters) depth to begin refraction calculations
+	float refract_step = 0.2;                  // (meters) step size to iterate through depth
+	float k_break = 0.5;                       // coefficient such that waves break at Hs > k_break*depth
 
 	int i = 0;
 	while (i < grid->num_shorelines)
@@ -31,8 +31,8 @@ void WaveTransformation(struct BeachGrid* grid, double wave_angle, double wave_p
 		do {
 			curr->transport_potential = 0;
 
-			double alpha_deep;
-			double shore_angle = (*grid).GetAngleByDifferencingScheme(grid, curr, wave_angle);
+			float alpha_deep;
+			float shore_angle = (*grid).GetAngleByDifferencingScheme(grid, curr, wave_angle);
 			alpha_deep = wave_angle - shore_angle;
 
 			if (fabs(alpha_deep) > (0.995 * PI / 2))
@@ -41,24 +41,24 @@ void WaveTransformation(struct BeachGrid* grid, double wave_angle, double wave_p
 				continue;
 			}
 
-			double local_wave_height = wave_height;
-			double c_deep = (GRAVITY * wave_period) / (2 * PI);
-			double l_deep = c_deep * wave_period;
-			double local_depth = start_depth + refract_step;
-			double local_alpha;
+			float local_wave_height = wave_height;
+			float c_deep = (GRAVITY * wave_period) / (2 * PI);
+			float l_deep = c_deep * wave_period;
+			float local_depth = start_depth + refract_step;
+			float local_alpha;
 
 			do {
 				// iterate
 				local_depth -= refract_step;
 
 				// non-iterative eqn or L, from Fenton & McKee
-				double wave_length = l_deep * powf(tanh(powf(powf(2.0 * PI / wave_period, 2.0) * (local_depth / GRAVITY), .75)), 2.0 / 3.0);
-				double local_c = wave_length / wave_period;
+				float wave_length = l_deep * powf(tanh(powf(powf(2.0 * PI / wave_period, 2.0) * (local_depth / GRAVITY), .75)), 2.0 / 3.0);
+				float local_c = wave_length / wave_period;
 
 				// n = 1/2(1+2kh/sinh(kh)) Komar 5.21
 				// kh = 2 pi depth/L  from k = 2 pi/L
-				double kh = 2 * PI * local_depth / wave_length;
-				double n = 0.5 * (1 + 2.0 * kh / sinh(2.0 * kh));
+				float kh = 2 * PI * local_depth / wave_length;
+				float n = 0.5 * (1 + 2.0 * kh / sinh(2.0 * kh));
 
 				// Calculate angle, assuming shore parallel contours and no conv/div of rays from Komar 5.47
 				local_alpha = asin(local_c / c_deep * sin(alpha_deep));
@@ -77,15 +77,15 @@ void WaveTransformation(struct BeachGrid* grid, double wave_angle, double wave_p
 	}
 }
 
-double GetTransportVolumePotential(double alpha, double wave_height, double timestep_length)
+float GetTransportVolumePotential(float alpha, float wave_height, float timestep_length)
 {
 	int rho = 1020;         // (kg/m^3) density of salt water
 	return fabs(1.1 * rho * powf(GRAVITY, 3.0 / 2.0) * powf(wave_height, 5.0 / 2.0) * cos(alpha) * sin(alpha) * timestep_length);
 }
 
-void GetAvailableSupply(struct BeachGrid* grid, int ref_pos, double ref_depth, double shelf_slope, double shoreface_slope, double min_depth)
+void GetAvailableSupply(struct BeachGrid* grid, int ref_pos, float ref_depth, float shelf_slope, float shoreface_slope, float min_depth)
 {
-	double cell_area = grid->cell_width * grid->cell_length;
+	float cell_area = grid->cell_width * grid->cell_length;
 
 	int i = 0;
 	while (i < grid->num_shorelines)
@@ -95,8 +95,8 @@ void GetAvailableSupply(struct BeachGrid* grid, int ref_pos, double ref_depth, d
 		struct BeachNode* curr = start;
 
 		do {
-			double volume_needed_left = 0.0;
-			double volume_needed_right = 0.0;
+			float volume_needed_left = 0.0;
+			float volume_needed_right = 0.0;
 
 			struct BeachNode* prev = curr->prev;
 
@@ -116,10 +116,10 @@ void GetAvailableSupply(struct BeachGrid* grid, int ref_pos, double ref_depth, d
 				break;
 			}
 
-			double total_volume_needed = volume_needed_left + volume_needed_right;
-			double shore_angle = (*grid).GetNextAngle(grid, curr);
-			double depth = GetDepthOfClosure(curr, ref_pos, ref_depth, shelf_slope, shoreface_slope, shore_angle, min_depth, grid->cell_length);
-			double volume_available = curr->frac_full * cell_area * depth;
+			float total_volume_needed = volume_needed_left + volume_needed_right;
+			float shore_angle = (*grid).GetNextAngle(grid, curr);
+			float depth = GetDepthOfClosure(curr, ref_pos, ref_depth, shelf_slope, shoreface_slope, shore_angle, min_depth, grid->cell_length);
+			float volume_available = curr->frac_full * cell_area * depth;
 			struct BeachNode* node_behind = GetNodeInDir(grid, curr, GetDir(shore_angle));
 			if (!BeachNode.isEmpty(node_behind) && node_behind->frac_full >= 1.0)
 			{
@@ -160,8 +160,8 @@ void NetVolumeChange(struct BeachGrid* grid)
 		struct BeachNode* curr = start;
 
 		do {
-			double volume_in = 0.0;
-			double volume_out = 0.0;
+			float volume_in = 0.0;
+			float volume_out = 0.0;
 
 			struct BeachNode* prev = curr->prev;
 
@@ -190,11 +190,11 @@ void NetVolumeChange(struct BeachGrid* grid)
 	}
 }
 
-int TransportSediment(struct BeachGrid* grid, int ref_pos, double ref_depth, double shelf_slope, double shoreface_slope, double min_depth)
+int TransportSediment(struct BeachGrid* grid, int ref_pos, float ref_depth, float shelf_slope, float shoreface_slope, float min_depth)
 {
 
 	int FIND_BEACH_FLAG = FALSE;
-	double cell_area = grid->cell_width * grid->cell_length;
+	float cell_area = grid->cell_width * grid->cell_length;
 	int i = 0;
 	while (i < grid->num_shorelines)
 	{
@@ -202,9 +202,9 @@ int TransportSediment(struct BeachGrid* grid, int ref_pos, double ref_depth, dou
 		struct BeachNode* curr = start;
 
 		do {
-			double shore_angle = (*grid).GetNextAngle(grid, curr);
-			double depth = GetDepthOfClosure(curr, ref_pos, ref_depth, shelf_slope, shoreface_slope, shore_angle, min_depth, grid->cell_length);
-			double net_area_change = curr->net_volume_change / depth;
+			float shore_angle = (*grid).GetNextAngle(grid, curr);
+			float depth = GetDepthOfClosure(curr, ref_pos, ref_depth, shelf_slope, shoreface_slope, shore_angle, min_depth, grid->cell_length);
+			float net_area_change = curr->net_volume_change / depth;
 			curr->frac_full = curr->frac_full + net_area_change / cell_area;
 			if (curr->frac_full < -0.000001)
 			{
@@ -254,7 +254,7 @@ void OopsImEmpty(struct BeachGrid* grid, struct BeachNode* node)
 		}
 	}
 
-	double delta_fill = node->frac_full / num_cells;
+	float delta_fill = node->frac_full / num_cells;
 	for (i = 0; i < 4; i++)
 	{
 		if (!BeachNode.isEmpty(neighbors[i]) && neighbors[i]->frac_full > 0 && (none_full || neighbors[i]->frac_full >= 1))
@@ -308,7 +308,7 @@ void OopsImFull(struct BeachGrid* grid, struct BeachNode* node)
 		}
 	}
 
-	double delta_fill = (node->frac_full - 1) / num_cells;
+	float delta_fill = (node->frac_full - 1) / num_cells;
 	for (i = 0; i < 4; i++)
 	{
 		if (!BeachNode.isEmpty(neighbors[i]) && neighbors[i]->frac_full < 1 && (none_empty || neighbors[i]->frac_full <= 0))
@@ -370,7 +370,7 @@ int FixBeach(struct BeachGrid* grid)
 				// distribute to beach neighbors
 				if (num_cells > 0)
 				{
-					double delta_fill = curr->frac_full / num_cells;
+					float delta_fill = curr->frac_full / num_cells;
 					curr->frac_full = 0;
 					for (j = 0; j < 4; j++)
 					{
@@ -435,20 +435,20 @@ int FixBeach(struct BeachGrid* grid)
 
 
 /* ---- SEDIMENT TRANSPORT HELPERS ------- */
-double GetDepthOfClosure(struct BeachNode* node, int ref_pos, double shelf_depth_at_ref_pos, double shelf_slope, double shoreface_slope, double shore_angle, double min_shelf_depth_at_closure, int cell_length)
+float GetDepthOfClosure(struct BeachNode* node, int ref_pos, float shelf_depth_at_ref_pos, float shelf_slope, float shoreface_slope, float shore_angle, float min_shelf_depth_at_closure, int cell_length)
 {
 	int x = node->row;
 	// Eq 1
-	double local_shelf_depth = shelf_depth_at_ref_pos + ((ref_pos - x) * cell_length * shelf_slope);
+	float local_shelf_depth = shelf_depth_at_ref_pos + ((ref_pos - x) * cell_length * shelf_slope);
 
 	// Eq 2
-	double cross_shore_distance_to_closure = local_shelf_depth / (shoreface_slope - (cos(shore_angle) * shelf_slope));
+	float cross_shore_distance_to_closure = local_shelf_depth / (shoreface_slope - (cos(shore_angle) * shelf_slope));
 
 	// Eq 3
-	double cross_shore_pos_of_closure = x + cos(shore_angle) * cross_shore_distance_to_closure / cell_length;
+	float cross_shore_pos_of_closure = x + cos(shore_angle) * cross_shore_distance_to_closure / cell_length;
 
 	// Eq 4
-	double closure_depth = shelf_depth_at_ref_pos + (cross_shore_pos_of_closure - ref_pos) * cell_length * shelf_slope;
+	float closure_depth = shelf_depth_at_ref_pos + (cross_shore_pos_of_closure - ref_pos) * cell_length * shelf_slope;
 	if (closure_depth < min_shelf_depth_at_closure)
 	{
 		closure_depth = min_shelf_depth_at_closure;
@@ -457,7 +457,7 @@ double GetDepthOfClosure(struct BeachNode* node, int ref_pos, double shelf_depth
 	return closure_depth;
 }
 
-double ModTowardZero(double a, double m)
+float ModTowardZero(float a, float m)
 {
 	int sign = a / fabs(a);
 	a = fabs(a);
@@ -467,17 +467,17 @@ double ModTowardZero(double a, double m)
 	return a - m * c;
 }
 
-double RoundRadians(double angle, double round_to, double bias)
+float RoundRadians(float angle, float round_to, float bias)
 {
-	double m90 = ModTowardZero(angle, PI / 2);
-	double m180 = ModTowardZero(angle, PI);
+	float m90 = ModTowardZero(angle, PI / 2);
+	float m180 = ModTowardZero(angle, PI);
 
 	if (m90 != m180)
 	{
 		bias = (PI / 2) - bias;
 	}
 
-	double rounded_angle = angle - m90;
+	float rounded_angle = angle - m90;
 	if (fabs(m90) >= bias)
 	{
 		rounded_angle += (m90 / fabs(m90)) * (PI / 2);
@@ -485,13 +485,13 @@ double RoundRadians(double angle, double round_to, double bias)
 	return rounded_angle;
 }
 
-double GetDir(double shore_angle)
+float GetDir(float shore_angle)
 {
-	double shore_normal = shore_angle - PI / 2;
+	float shore_normal = shore_angle - PI / 2;
 	return RoundRadians(shore_normal, PI / 2, PI / 6);
 }
 
-struct BeachNode* GetNodeInDir(struct BeachGrid* grid, struct BeachNode* node, double dir)
+struct BeachNode* GetNodeInDir(struct BeachGrid* grid, struct BeachNode* node, float dir)
 {
 	int r = node->row;
 	int c = node->col;
